% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_structural_shocks.PosteriorBSVARSIGN}
\alias{compute_structural_shocks.PosteriorBSVARSIGN}
\title{Computes posterior draws of structural shocks}
\usage{
\method{compute_structural_shocks}{PosteriorBSVARSIGN}(posterior)
}
\arguments{
\item{posterior}{posterior estimation outcome - an object of class 
\code{PosteriorBSVARSIGN} obtained by running the \code{estimate} function.}
}
\value{
An object of class \code{PosteriorShocks}, that is, an \code{NxTxS} 
array with attribute \code{PosteriorShocks} containing \code{S} draws of the 
structural shocks.
}
\description{
Each of the draws from the posterior estimation of models from
packages \pkg{bsvars} or \pkg{bsvarSIGNs} is transformed into
a draw from the posterior distribution of the structural shocks.
}
\examples{
# upload data
data(optimism)

# specify the model and set seed
set.seed(123)
sign_irf       = matrix(c(0, 1, rep(NA, 23)), 5, 5)
specification  = specify_bsvarSIGN$new(optimism, sign_irf = sign_irf)

# run the burn-in
posterior      = estimate(specification, 10)

# compute structural shocks
shocks         = compute_structural_shocks(posterior)

# workflow with the pipe |>
############################################################
set.seed(123)
optimism |>
  specify_bsvarSIGN$new(sign_irf = sign_irf) |> 
  estimate(S = 20) |> 
  compute_structural_shocks() -> ss

}
\seealso{
\code{\link{estimate.BSVARSIGN}}, \code{\link{summary}}, \code{\link{plot}}
}
\author{
Xiaolei Wang \email{adamwang15@gmail.com} and Tomasz Woźniak \email{wozniak.tom@pm.me}
}
