% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_par.R
\name{estimate_par}
\alias{estimate_par}
\title{Parallel Bayesian estimation of a Structural Vector Autoregression
with traditional and narrative sign restrictions}
\usage{
estimate_par(
  specification,
  S,
  thin = 1,
  show_progress = TRUE,
  ncores = parallel::detectCores() - 1
)
}
\arguments{
\item{specification}{an object of class BSVARSIGN generated using the \code{specify_bsvarSIGN$new()} function.}

\item{S}{a positive integer, the number of posterior draws to be generated}

\item{thin}{a positive integer, specifying the frequency of MCMC output thinning}

\item{show_progress}{a logical value, if \code{TRUE} the estimation progress bar is visible}

\item{ncores}{a positive integer, the number of cores to be used for parallel execution}
}
\value{
An object of class \code{PosteriorBSVARSIGN}
}
\description{
Parallelised version of \code{estimate} using the \code{parallel} package.
Estimates Bayesian Structural Vector Autoregression model
using independent sampling.
}
