% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{importance_sampling}
\alias{importance_sampling}
\title{Importance sampling}
\usage{
importance_sampling(posterior)
}
\arguments{
\item{posterior}{posterior estimation outcome}
}
\description{
Given posterior draws with importance weights, sample with replacement
}
