% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oil}
\alias{oil}
\title{A 3-variable oil price data, from 1971 to 2016}
\format{
A matrix and a \code{ts} object with time series of over five hundred observations on 3 variables:
\describe{
  \item{opg}{monthly growth rate of global oil production}
  \item{eai}{monthly index of real economic activity}
  \item{rop}{monthly real oil price}
}

The series are as described by Antolín-Díaz & Rubio-Ramírez (2018) in section IV-A.
}
\source{
Replication package, \url{https://www.aeaweb.org/articles?id=10.1257/aer.20161852}
}
\usage{
data(oil)
}
\description{
A sample data to identify oil demand and supply shocks.
}
\references{
Antolín-Díaz & Rubio-Ramírez (2018) Narrative Sign Restrictions for SVARs, American Economic Review, 108(10), 2802-29, <doi:10.1257/aer.20161852>.
}
\keyword{datasets}
