% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_bsvarSIGN.R
\name{specify_data_matrices}
\alias{specify_data_matrices}
\title{R6 Class Representing DataMatricesBSVARSIGN}
\description{
The class DataMatricesBSVARSIGN presents the data matrices of dependent variables, \eqn{Y}, 
and regressors, \eqn{X}, for the homoskedastic bsvar model.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Y}}{an \code{NxT} matrix of dependent variables, \eqn{Y}.}

\item{\code{X}}{an \code{KxT} matrix of regressors, \eqn{X}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataMatricesBSVARSIGN-new}{\code{specify_data_matrices$new()}}
\item \href{#method-DataMatricesBSVARSIGN-get_data_matrices}{\code{specify_data_matrices$get_data_matrices()}}
\item \href{#method-DataMatricesBSVARSIGN-clone}{\code{specify_data_matrices$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataMatricesBSVARSIGN-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataMatricesBSVARSIGN-new}{}}}
\subsection{Method \code{new()}}{
Create new data matrices DataMatricesBSVARSIGN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_data_matrices$new(data, p = 1L, exogenous = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a \code{(T+p)xN} matrix with time series data.}

\item{\code{p}}{a positive integer providing model's autoregressive lag order.}

\item{\code{exogenous}}{a \code{(T+p)xd} matrix of exogenous variables. 
This matrix should not include a constant term.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
New data matrices DataMatricesBSVARSIGN.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataMatricesBSVARSIGN-get_data_matrices"></a>}}
\if{latex}{\out{\hypertarget{method-DataMatricesBSVARSIGN-get_data_matrices}{}}}
\subsection{Method \code{get_data_matrices()}}{
Returns the data matrices DataMatricesBSVARSIGN as a \code{list}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_data_matrices$get_data_matrices()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataMatricesBSVARSIGN-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataMatricesBSVARSIGN-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_data_matrices$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
