% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_bsvarSIGN.R
\name{specify_identification_bsvarSIGNs}
\alias{specify_identification_bsvarSIGNs}
\title{R6 Class Representing IdentificationBSVARSIGNs}
\description{
The class IdentificationBSVARSIGNs presents the identifying restrictions for the bsvar models.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{VB}}{a list of \code{N} matrices determining the unrestricted elements of matrix \eqn{B}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IdentificationBSVARSIGNs-new}{\code{specify_identification_bsvarSIGNs$new()}}
\item \href{#method-IdentificationBSVARSIGNs-get_identification}{\code{specify_identification_bsvarSIGNs$get_identification()}}
\item \href{#method-IdentificationBSVARSIGNs-set_identification}{\code{specify_identification_bsvarSIGNs$set_identification()}}
\item \href{#method-IdentificationBSVARSIGNs-clone}{\code{specify_identification_bsvarSIGNs$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentificationBSVARSIGNs-new"></a>}}
\if{latex}{\out{\hypertarget{method-IdentificationBSVARSIGNs-new}{}}}
\subsection{Method \code{new()}}{
Create new identifying restrictions IdentificationBSVARSIGNs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_identification_bsvarSIGNs$new(N, B)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{a positive integer - the number of dependent variables in the model.}

\item{\code{B}}{a logical \code{NxN} matrix containing value \code{TRUE} for the elements of 
the structural matrix \eqn{B} to be estimated and value \code{FALSE} for exclusion restrictions 
to be set to zero.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Identifying restrictions IdentificationBSVARSIGNs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentificationBSVARSIGNs-get_identification"></a>}}
\if{latex}{\out{\hypertarget{method-IdentificationBSVARSIGNs-get_identification}{}}}
\subsection{Method \code{get_identification()}}{
Returns the elements of the identification pattern IdentificationBSVARSIGNs as a \code{list}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_identification_bsvarSIGNs$get_identification()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentificationBSVARSIGNs-set_identification"></a>}}
\if{latex}{\out{\hypertarget{method-IdentificationBSVARSIGNs-set_identification}{}}}
\subsection{Method \code{set_identification()}}{
Set new starting values StartingValuesBSVARSIGN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_identification_bsvarSIGNs$set_identification(N, B)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{a positive integer - the number of dependent variables in the model.}

\item{\code{B}}{a logical \code{NxN} matrix containing value \code{TRUE} for the elements of 
the structural matrix \eqn{B} to be estimated and value \code{FALSE} for exclusion restrictions 
to be set to zero.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentificationBSVARSIGNs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IdentificationBSVARSIGNs-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_identification_bsvarSIGNs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
