% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify_narrative}
\alias{specify_narrative}
\title{vector specifying one narrative restriction}
\usage{
specify_narrative(
  start,
  periods = 1,
  type = "S",
  sign = 1,
  shock = 1,
  var = NA
)
}
\description{
The class narrative specifies a single narrative restriction.
}
\examples{
# a prior for 5-variable example with one lag 
specify_narrative(start = 166, periods = 1, type = "S", sign = 1, shock = 1, var = 6)

}
