% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_bsvarSIGN.R
\name{specify_starting_values_bsvarSIGN}
\alias{specify_starting_values_bsvarSIGN}
\title{R6 Class Representing StartingValuesBSVARSIGN}
\description{
The class StartingValuesBSVARSIGN presents starting values for the homoskedastic bsvar model.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{A}}{an \code{NxK} matrix of starting values for the parameter \eqn{A}.}

\item{\code{B}}{an \code{NxN} matrix of starting values for the parameter \eqn{B}.}

\item{\code{hyper}}{a \code{(2*N+1)x2} matrix of starting values for the shrinkage hyper-parameters of the 
hierarchical prior distribution.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StartingValuesBSVARSIGN-new}{\code{specify_starting_values_bsvarSIGN$new()}}
\item \href{#method-StartingValuesBSVARSIGN-get_starting_values}{\code{specify_starting_values_bsvarSIGN$get_starting_values()}}
\item \href{#method-StartingValuesBSVARSIGN-set_starting_values}{\code{specify_starting_values_bsvarSIGN$set_starting_values()}}
\item \href{#method-StartingValuesBSVARSIGN-clone}{\code{specify_starting_values_bsvarSIGN$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartingValuesBSVARSIGN-new"></a>}}
\if{latex}{\out{\hypertarget{method-StartingValuesBSVARSIGN-new}{}}}
\subsection{Method \code{new()}}{
Create new starting values StartingValuesBSVARSIGN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_starting_values_bsvarSIGN$new(N, p, d = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{a positive integer - the number of dependent variables in the model.}

\item{\code{p}}{a positive integer - the autoregressive lag order of the SVAR model.}

\item{\code{d}}{a positive integer - the number of \code{exogenous} variables in the model.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Starting values StartingValuesBSVARSIGN.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartingValuesBSVARSIGN-get_starting_values"></a>}}
\if{latex}{\out{\hypertarget{method-StartingValuesBSVARSIGN-get_starting_values}{}}}
\subsection{Method \code{get_starting_values()}}{
Returns the elements of the starting values StartingValuesBSVARSIGN as a \code{list}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_starting_values_bsvarSIGN$get_starting_values()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartingValuesBSVARSIGN-set_starting_values"></a>}}
\if{latex}{\out{\hypertarget{method-StartingValuesBSVARSIGN-set_starting_values}{}}}
\subsection{Method \code{set_starting_values()}}{
Returns the elements of the starting values StartingValuesBSVARSIGN as a \code{list}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_starting_values_bsvarSIGN$set_starting_values(last_draw)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{last_draw}}{a list containing the last draw of elements \code{B} - an \code{NxN} matrix, 
\code{A} - an \code{NxK} matrix, and \code{hyper} - a vector of 5 positive real numbers.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class StartingValuesBSVARSIGN including the last draw of the current MCMC 
as the starting value to be passed to the continuation of the MCMC estimation using \code{estimate()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StartingValuesBSVARSIGN-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StartingValuesBSVARSIGN-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_starting_values_bsvarSIGN$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
