\name{bsvarSIGNs-package}
\alias{bsvarSIGNs-package}
\alias{bsvarSIGNs}
\docType{package}
\title{
\packageTitle{bsvarSIGNs}
}
\description{
\packageDescription{bsvarSIGNs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bsvarSIGNs}
\packageIndices{bsvarSIGNs}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{bsvarSIGNs}

Maintainer: \packageMaintainer{bsvarSIGNs}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
