% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_bsvarSIGN.R
\name{specify_bsvarSIGN}
\alias{specify_bsvarSIGN}
\title{R6 Class representing the specification of the homoskedastic BSVARSIGN model}
\description{
The class BSVARSIGN presents complete specification for the homoskedastic bsvar model.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{p}}{a non-negative integer specifying the autoregressive lag order of the model.}

\item{\code{identification}}{an object IdentificationBSVARSIGN with the identifying restrictions.}

\item{\code{prior}}{an object PriorBSVARSIGN with the prior specification.}

\item{\code{data_matrices}}{an object DataMatricesBSVARSIGN with the data matrices.}

\item{\code{starting_values}}{an object StartingValuesBSVARSIGN with the starting values.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BSVARSIGN-new}{\code{specify_bsvarSIGN$new()}}
\item \href{#method-BSVARSIGN-get_data_matrices}{\code{specify_bsvarSIGN$get_data_matrices()}}
\item \href{#method-BSVARSIGN-get_identification}{\code{specify_bsvarSIGN$get_identification()}}
\item \href{#method-BSVARSIGN-get_prior}{\code{specify_bsvarSIGN$get_prior()}}
\item \href{#method-BSVARSIGN-get_starting_values}{\code{specify_bsvarSIGN$get_starting_values()}}
\item \href{#method-BSVARSIGN-clone}{\code{specify_bsvarSIGN$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BSVARSIGN-new"></a>}}
\if{latex}{\out{\hypertarget{method-BSVARSIGN-new}{}}}
\subsection{Method \code{new()}}{
Create a new specification of the homoskedastic bsvar model BSVARSIGN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_bsvarSIGN$new(
  data,
  p = 1L,
  sign_irf,
  sign_hd,
  exogenous = NULL,
  stationary = rep(FALSE, ncol(data))
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a \code{(T+p)xN} matrix with time series data.}

\item{\code{p}}{a positive integer providing model's autoregressive lag order.}

\item{\code{sign_irf}}{a \code{NxNxh} array of sign restrictions on the impulse response functions
up to \code{h} horizons. Each \code{NxN} matrix nnly accept values in {-1 ,0, 1},
-1: negative restriction, 0: no restriction, 1: positive restriction.}

\item{\code{sign_hd}}{a \code{Mx6} matrix of sign restrictions on the historical decompositions.
Each row of the matrix corresponds to a different restriction.
Each column contains the following information:
| type | variable | shock | start | end | sign |}

\item{\code{exogenous}}{a \code{(T+p)xd} matrix of exogenous variables.}

\item{\code{stationary}}{an \code{N} logical vector - its element set to \code{FALSE} sets
the prior mean for the autoregressive parameters of the \code{N}th equation to the white noise process,
otherwise to random walk.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new complete specification for the homoskedastic bsvar model BSVARSIGN.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BSVARSIGN-get_data_matrices"></a>}}
\if{latex}{\out{\hypertarget{method-BSVARSIGN-get_data_matrices}{}}}
\subsection{Method \code{get_data_matrices()}}{
Returns the data matrices as the DataMatricesBSVARSIGN object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_bsvarSIGN$get_data_matrices()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BSVARSIGN-get_identification"></a>}}
\if{latex}{\out{\hypertarget{method-BSVARSIGN-get_identification}{}}}
\subsection{Method \code{get_identification()}}{
Returns the identifying restrictions as the IdentificationBSVARSIGNs object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_bsvarSIGN$get_identification()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BSVARSIGN-get_prior"></a>}}
\if{latex}{\out{\hypertarget{method-BSVARSIGN-get_prior}{}}}
\subsection{Method \code{get_prior()}}{
Returns the prior specification as the PriorBSVARSIGN object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_bsvarSIGN$get_prior()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BSVARSIGN-get_starting_values"></a>}}
\if{latex}{\out{\hypertarget{method-BSVARSIGN-get_starting_values}{}}}
\subsection{Method \code{get_starting_values()}}{
Returns the starting values as the StartingValuesBSVARSIGN object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_bsvarSIGN$get_starting_values()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BSVARSIGN-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BSVARSIGN-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_bsvarSIGN$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
