% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_bsvarSIGN.R
\name{specify_identification_bsvarSIGN}
\alias{specify_identification_bsvarSIGN}
\title{R6 Class Representing IdentificationBSVARSIGNs}
\description{
The class IdentificationBSVARSIGNs presents the identifying restrictions for the bsvar models.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{VB}}{a list of \code{N} matrices determining the unrestricted elements of matrix \eqn{B}.}

\item{\code{sign_irf}}{a \code{NxNxh} array of sign restrictions on the impulse response functions.}

\item{\code{sign_narrative}}{a \code{Kx6} matrix of narrative sign restrictions.}

\item{\code{sign_B}}{a \code{NxN} matrix of sign restrictions on contemporaneous relations.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IdentificationBSVARSIGNs-new}{\code{specify_identification_bsvarSIGN$new()}}
\item \href{#method-IdentificationBSVARSIGNs-get_identification}{\code{specify_identification_bsvarSIGN$get_identification()}}
\item \href{#method-IdentificationBSVARSIGNs-set_identification}{\code{specify_identification_bsvarSIGN$set_identification()}}
\item \href{#method-IdentificationBSVARSIGNs-clone}{\code{specify_identification_bsvarSIGN$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentificationBSVARSIGNs-new"></a>}}
\if{latex}{\out{\hypertarget{method-IdentificationBSVARSIGNs-new}{}}}
\subsection{Method \code{new()}}{
Create new identifying restrictions IdentificationBSVARSIGNs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_identification_bsvarSIGN$new(N, sign_irf, sign_narrative, sign_B)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{a positive integer - the number of dependent variables in the model.}

\item{\code{sign_irf}}{a \code{NxNxh} array of sign restrictions on the impulse response functions
up to \code{h} horizons. Each \code{NxN} matrix only accept values in {-1 ,0, 1},
-1: negative restriction, 0: no restriction, 1: positive restriction.}

\item{\code{sign_narrative}}{a \code{Kx6} matrix of narrative sign restrictions,
each row of the matrix corresponds to a different restriction,
each column contains the following information: \cr
| type | sign | var_i | shock_j | start | horizons | \cr
type: 0 if no restriction; \cr
1 if restriction on structural shock; \cr
2 if type A restriction on historical decomposition
i.e. hd of shock_j on var_i is greater (less) than 0; \cr
3 if type B restriction on historical decomposition
i.e. hd of shock_j on var_i is the largest (smallest); \cr
sign: depending on type, 1 if greater/largest, otherwise less/smallest; \cr
var_i: \code{i}-th variable; \cr
shock_j: \code{j}-th shock; \cr
start: starting period of the restriction; \cr
horizons: horizons of the restriction, input 0 if only one period \cr}

\item{\code{sign_B}}{a \code{NxN} matrix of sign restrictions on contemporaneous relations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Identifying restrictions IdentificationBSVARSIGNs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentificationBSVARSIGNs-get_identification"></a>}}
\if{latex}{\out{\hypertarget{method-IdentificationBSVARSIGNs-get_identification}{}}}
\subsection{Method \code{get_identification()}}{
Returns the elements of the identification pattern IdentificationBSVARSIGNs as a \code{list}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_identification_bsvarSIGN$get_identification()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentificationBSVARSIGNs-set_identification"></a>}}
\if{latex}{\out{\hypertarget{method-IdentificationBSVARSIGNs-set_identification}{}}}
\subsection{Method \code{set_identification()}}{
Set new starting values StartingValuesBSVARSIGN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_identification_bsvarSIGN$set_identification(
  N,
  sign_irf,
  sign_narrative,
  sign_B
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{a positive integer - the number of dependent variables in the model.}

\item{\code{sign_irf}}{a \code{NxNxh} array of sign restrictions on the impulse response functions
up to \code{h} horizons. Each \code{NxN} matrix only accept values in {-1 ,0, 1},
-1: negative restriction, 0: no restriction, 1: positive restriction.}

\item{\code{sign_narrative}}{a \code{Mx6} matrix of narrative sign restrictions,
each row of the matrix corresponds to a different restriction,
each column contains the following information: \cr
| type | sign | var_i | shock_j | start | horizons | \cr
type: 0 if no restriction; \cr
1 if restriction on structural shock; \cr
2 if type A restriction on historical decomposition
i.e. hd of shock_j on var_i is greater (less) than 0; \cr
3 if type B restriction on historical decomposition
i.e. hd of shock_j on var_i is the largest (smallest); \cr
sign: depending on type, 1 if greater/largest, otherwise less/smallest; \cr
var_i: \code{i}-th variable; \cr
shock_j: \code{j}-th shock; \cr
start: starting period of the restriction; \cr
horizons: horizons of the restriction, input 0 if only one period \cr}

\item{\code{sign_B}}{a \code{NxN} matrix of sign restrictions on contemporaneous relations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentificationBSVARSIGNs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IdentificationBSVARSIGNs-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_identification_bsvarSIGN$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
