% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_bsvarSIGN.R
\name{specify_identification_bsvarSIGN}
\alias{specify_identification_bsvarSIGN}
\title{R6 Class Representing IdentificationBSVARSIGN}
\description{
The class IdentificationBSVARSIGN presents the identifying restrictions for the Bayesian Structural VAR models with sign and narrative restrictions.
}
\examples{
specify_identification_bsvarSIGN$new(N = 5) # recursive specification for a 5-variable system

# an identification pattern with narrative sign restrictions
sign_irf = matrix(c(0, 1, rep(NA, 23)), 5, 5)
specify_identification_bsvarSIGN$new(N = 5, sign_irf = sign_irf) 

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{VB}}{a list of \code{N} matrices determining the unrestricted elements of matrix \eqn{B}.}

\item{\code{sign_irf}}{a \code{NxNxH} array of sign restrictions on the impulse response functions.}

\item{\code{sign_narrative}}{a \code{Mx6} matrix of narrative sign restrictions.}

\item{\code{sign_relation}}{a \code{NxN} matrix of sign restrictions on contemporaneous relations.}

\item{\code{max_tries}}{a positive integer with the maximum number of iterations 
for finding a rotation matrix \eqn{Q} that would satisfy sign restrictions.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IdentificationBSVARSIGN-new}{\code{specify_identification_bsvarSIGN$new()}}
\item \href{#method-IdentificationBSVARSIGN-get_identification}{\code{specify_identification_bsvarSIGN$get_identification()}}
\item \href{#method-IdentificationBSVARSIGN-set_identification}{\code{specify_identification_bsvarSIGN$set_identification()}}
\item \href{#method-IdentificationBSVARSIGN-clone}{\code{specify_identification_bsvarSIGN$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentificationBSVARSIGN-new"></a>}}
\if{latex}{\out{\hypertarget{method-IdentificationBSVARSIGN-new}{}}}
\subsection{Method \code{new()}}{
Create new identifying restrictions IdentificationBSVARSIGN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_identification_bsvarSIGN$new(
  N,
  sign_irf,
  sign_narrative,
  sign_relation,
  max_tries = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{a positive integer - the number of dependent variables in the model.}

\item{\code{sign_irf}}{a \code{NxNxH} array with entries in (-1 ,0, 1), sign restrictions on the
impulse response functions, the \code{h}-th slice \code{NxN} matrix contains the
sign restrictions on the \code{h-1} horizon, e.g. \code{sign_irf[,,0]} contains restrictions
on the contemporaneous impulse response function.}

\item{\code{sign_narrative}}{a \code{Mx6} matrix of narrative sign restrictions,
each row of the matrix corresponds to a different restriction,
columns are (type, sign, var_i, shock_j, start_t, horizons_h) with detailed definitions: \cr\cr
Column 1 (type):
0 if no restriction;
1 if restriction on structural shock;
2 if type A restriction on historical decomposition
i.e. historical decomposition of shock_j on var_i is greater (less) than 0;
3 if type B restriction on historical decomposition
i.e. historical decomposition of shock_j on var_i is the largest (smallest); \cr
Column 2 (sign): depending on type, 1 if greater/largest, -1 if less/smallest. \cr
Column 3 (var_i): an integer in 1:N (or NA when type = 0), index of the restricted variable. \cr
Column 4 (shock_j): an integer in 1:N, index of the restricted shock. \cr
Column 5 (start_t): an integer in 1:T, starting period of the restriction; \cr
Column 6 (horizons_h): an integer in 1:(T-start_t), number horizons of the restriction,
if start=t and horizons=h the restriction in on periods t to t+h,
e.g. when h=0 the restriction in only placed on period t.}

\item{\code{sign_relation}}{a \code{NxN} matrix with entries in (-1 ,0, 1), sign restrictions on the
contemporaneous relations \code{B} between reduced-form errors \code{E} and
structural shocks \code{U}. Recall the structural equation \code{BE=U}, the inverse
of \code{B} is the contemporaneous impulse response function.}

\item{\code{max_tries}}{a positive integer with the maximum number of iterations
for finding a rotation matrix \eqn{Q} that would satisfy sign restrictions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Identifying restrictions IdentificationBSVARSIGN.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentificationBSVARSIGN-get_identification"></a>}}
\if{latex}{\out{\hypertarget{method-IdentificationBSVARSIGN-get_identification}{}}}
\subsection{Method \code{get_identification()}}{
Returns the elements of the identification pattern IdentificationBSVARSIGN as a \code{list}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_identification_bsvarSIGN$get_identification()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentificationBSVARSIGN-set_identification"></a>}}
\if{latex}{\out{\hypertarget{method-IdentificationBSVARSIGN-set_identification}{}}}
\subsection{Method \code{set_identification()}}{
Set new starting values StartingValuesBSVARSIGN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_identification_bsvarSIGN$set_identification(
  N,
  sign_irf,
  sign_narrative,
  sign_relation
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{N}}{a positive integer - the number of dependent variables in the model.}

\item{\code{sign_irf}}{a \code{NxNxH} array with entries in (-1 ,0, 1), sign restrictions on the
impulse response functions, the \code{h}-th slice \code{NxN} matrix contains the
sign restrictions on the \code{h-1} horizon, e.g. \code{sign_irf[,,0]} contains restrictions
on the contemporaneous impulse response function.}

\item{\code{sign_narrative}}{a \code{Mx6} matrix of narrative sign restrictions,
each row of the matrix corresponds to a different restriction,
columns are (type, sign, var_i, shock_j, start_t, horizons_h) with detailed definitions: \cr\cr
Column 1 (type):
0 if no restriction;
1 if restriction on structural shock;
2 if type A restriction on historical decomposition
i.e. historical decomposition of shock_j on var_i is greater (less) than 0;
3 if type B restriction on historical decomposition
i.e. historical decomposition of shock_j on var_i is the largest (smallest); \cr
Column 2 (sign): depending on type, 1 if greater/largest, -1 if less/smallest. \cr
Column 3 (var_i): an integer in 1:N (or NA when type = 0), index of the restricted variable. \cr
Column 4 (shock_j): an integer in 1:N, index of the restricted shock. \cr
Column 5 (start_t): an integer in 1:T, starting period of the restriction; \cr
Column 6 (horizons_h): an integer in 1:(T-start_t), number horizons of the restriction,
if start=t and horizons=h the restriction in on periods t to t+h,
e.g. when h=0 the restriction in only placed on period t.}

\item{\code{sign_relation}}{a \code{NxN} matrix with entries in (-1 ,0, 1), sign restrictions on the
contemporaneous relations \code{B} between reduced-form errors \code{E} and
structural shocks \code{U}. Recall the structural equation \code{BE=U}, the inverse
of \code{B} is the contemporaneous impulse response function.}

\item{\code{max_tries}}{a positive integer with the maximum number of iterations
for finding a rotation matrix \eqn{Q} that would satisfy sign restrictions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentificationBSVARSIGN-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IdentificationBSVARSIGN-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specify_identification_bsvarSIGN$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
